Updated test parser with problem parsing. All the code is there but only domain parses. All tests pass.
