+++ lib/parsing/domain.cpp
@@ -4,5 +4,6 @@
 namespace client {
     namespace parser {
         BOOST_SPIRIT_INSTANTIATE(domain_type, iterator_type, context_type);
+        BOOST_SPIRIT_INSTANTIATE(problem_type, iterator_type, context_type);
     }
 } // namespace client
