+++ test/CMakeLists.txt
@@ -1,9 +1,12 @@
+#cmake_minimum_required(VERSION 3.19)
+
 set(CMAKE_CXX_STANDARD 20)
 
 find_package(Boost REQUIRED)
+
 add_executable(test_parser test_parser.cpp)
-target_include_directories(test_parser PRIVATE ${Boost_INCLUDE_DIR} ../lib)
 target_link_libraries(test_parser tomcat)
+target_include_directories(test_parser PRIVATE ${Boost_INCLUDE_DIR} ../lib)
 
 # KB testing
 add_executable(test_kb test_kb.cpp)
