+++ test/test_problem2.pddl
@@ -1,0 +1,8 @@
+(define
+    (problem bring-home-bacon)
+    (:domain construction)
+    (:requirements :strips)
+    (:objects
+        work house - site
+        tocino - material)
+)
